(FM-7727) Update to use RSAPI Transports
