+++ Gemfile
@@ -27,7 +27,6 @@ group :development do
   gem "puppet-module-win-default-r#{minor_version}",   require: false, platforms: [:mswin, :mingw, :x64_mingw]
   gem "puppet-module-win-dev-r#{minor_version}",       require: false, platforms: [:mswin, :mingw, :x64_mingw]
   gem "codecov",                                       require: false
-  gem "pry-byebug",                                    require: false
   gem "puppet-resource_api",                           require: false
   gem "simplecov-console",                             require: false
   gem "github_changelog_generator",                    require: false, git: 'https://github.com/skywinder/github-changelog-generator', ref: 'master' if Gem::Version.new(RUBY_VERSION.dup) >= Gem::Version.new('2.2.2')
