+++ lib/puppet/provider/ios_aaa_accounting/cisco_ios.rb
@@ -48,7 +48,7 @@ class Puppet::Provider::IosAaaAccounting::CiscoIos
   end
 
   def get(context)
-    output = context.device.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
+    output = context.transport.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
     return [] if output.nil?
     return_value = Puppet::Provider::IosAaaAccounting::CiscoIos.instances_from_cli(output)
     PuppetX::CiscoIOS::Utility.enforce_simple_types(context, return_value)
@@ -73,7 +73,7 @@ class Puppet::Provider::IosAaaAccounting::CiscoIos
   def update(context, _name, should)
     array_of_commands_to_run = Puppet::Provider::IosAaaAccounting::CiscoIos.commands_from_instance(should)
     array_of_commands_to_run.each do |command|
-      context.device.run_command_conf_t_mode(command)
+      context.transport.run_command_conf_t_mode(command)
     end
   end
 
@@ -81,7 +81,7 @@ class Puppet::Provider::IosAaaAccounting::CiscoIos
     is[:ensure] = 'absent'
     array_of_commands_to_run = Puppet::Provider::IosAaaAccounting::CiscoIos.commands_from_instance(is)
     array_of_commands_to_run.each do |command|
-      context.device.run_command_conf_t_mode(command)
+      context.transport.run_command_conf_t_mode(command)
     end
   end
 end
