+++ lib/puppet/provider/ios_access_list/ios.rb
@@ -37,7 +37,7 @@ class Puppet::Provider::IosAccessList::IosAccessList
   end
 
   def get(context)
-    output = context.device.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
+    output = context.transport.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
     return [] if output.nil?
     Puppet::Provider::IosAccessList::IosAccessList.instances_from_cli(output)
   end
@@ -61,7 +61,7 @@ class Puppet::Provider::IosAccessList::IosAccessList
   def update(context, _name, should)
     array_of_commands_to_run = Puppet::Provider::IosAccessList::IosAccessList.commands_from_instance(should)
     array_of_commands_to_run.each do |command|
-      context.device.run_command_conf_t_mode(command)
+      context.transport.run_command_conf_t_mode(command)
     end
   end
 
@@ -69,7 +69,7 @@ class Puppet::Provider::IosAccessList::IosAccessList
     is[:ensure] = 'absent'
     array_of_commands_to_run = Puppet::Provider::IosAccessList::IosAccessList.commands_from_instance(is)
     array_of_commands_to_run.each do |command|
-      context.device.run_command_conf_t_mode(command)
+      context.transport.run_command_conf_t_mode(command)
     end
   end
 end
