+++ lib/puppet/provider/ios_stp_global/cisco_ios.rb
@@ -66,7 +66,7 @@ class Puppet::Provider::IosStpGlobal::CiscoIos
   end
 
   def get(context)
-    output = context.device.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
+    output = context.transport.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
     return [] if output.nil?
     return_value = Puppet::Provider::IosStpGlobal::CiscoIos.instances_from_cli(output)
     PuppetX::CiscoIOS::Utility.enforce_simple_types(context, return_value)
@@ -91,11 +91,11 @@ class Puppet::Provider::IosStpGlobal::CiscoIos
   def update(context, _name, should)
     array_of_commands_mst_mode = Puppet::Provider::IosStpGlobal::CiscoIos.mst_commands_from_instance(should)
     array_of_commands_mst_mode.each do |command|
-      context.device.run_command_mst_mode(command)
+      context.transport.run_command_mst_mode(command)
     end
     array_of_commands_to_run = Puppet::Provider::IosStpGlobal::CiscoIos.commands_from_instance(should)
     array_of_commands_to_run.each do |command|
-      context.device.run_command_conf_t_mode(command)
+      context.transport.run_command_conf_t_mode(command)
     end
   end
 
@@ -103,12 +103,12 @@ class Puppet::Provider::IosStpGlobal::CiscoIos
     array_of_commands_mst_mode = Puppet::Provider::IosStpGlobal::CiscoIos.mst_commands_from_instance(is)
     array_of_commands_mst_mode.each do |command|
       command = 'no ' + command
-      context.device.run_command_mst_mode(command)
+      context.transport.run_command_mst_mode(command)
     end
     array_of_commands_to_run = Puppet::Provider::IosStpGlobal::CiscoIos.commands_from_instance(is)
     array_of_commands_to_run.each do |command|
       command = 'no ' + command
-      context.device.run_command_conf_t_mode(command)
+      context.transport.run_command_conf_t_mode(command)
     end
   end
 end
