+++ lib/puppet/provider/network_trunk/cisco_ios.rb
@@ -64,13 +64,13 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     end
 
     def get(context, _names = nil)
-      name_output = context.device.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_interface_names(commands_hash))
+      name_output = context.transport.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_interface_names(commands_hash))
       interface_names = Puppet::Provider::NetworkTrunk::CiscoIos.interface_names_from_cli(name_output)
 
       return_instances = []
       [*interface_names].each do |interface_name|
         get_value_cmd = PuppetX::CiscoIOS::Utility.get_values(commands_hash).to_s.gsub(%r{<name>}, interface_name)
-        output = context.device.run_command_enable_mode(get_value_cmd)
+        output = context.transport.run_command_enable_mode(get_value_cmd)
         # If this interface is not a switchable port ignore
         if !output.nil? && (!output.include? ' is not a switchable port')
           return_instances << Puppet::Provider::NetworkTrunk::CiscoIos.instance_from_cli(output, interface_name)
@@ -80,14 +80,14 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     end
 
     def create(context, name, should)
-      context.device.run_command_interface_mode(name, Puppet::Provider::NetworkTrunk::CiscoIos.commands_from_instance(should).join("\n"))
+      context.transport.run_command_interface_mode(name, Puppet::Provider::NetworkTrunk::CiscoIos.commands_from_instance(should).join("\n"))
     end
 
     alias update create
 
     def delete(context, name)
       delete_hash = { name: name, ensure: 'absent' }
-      context.device.run_command_interface_mode(name, Puppet::Provider::NetworkTrunk::CiscoIos.commands_from_instance(delete_hash).join("\n"))
+      context.transport.run_command_interface_mode(name, Puppet::Provider::NetworkTrunk::CiscoIos.commands_from_instance(delete_hash).join("\n"))
     end
 
     def canonicalize(_context, resources)
