+++ lib/puppet/provider/network_vlan/cisco_ios.rb
@@ -55,7 +55,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     end
 
     def get(context, _names = nil)
-      output = context.device.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
+      output = context.transport.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
       return [] if output.nil?
       return_value = Puppet::Provider::NetworkVlan::CiscoIos.instances_from_cli(output)
       PuppetX::CiscoIOS::Utility.enforce_simple_types(context, return_value)
@@ -64,7 +64,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     def update(context, name, should)
       array_of_commands_to_run = Puppet::Provider::NetworkVlan::CiscoIos.commands_from_instance(should)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_vlan_mode(name, command)
+        context.transport.run_command_vlan_mode(name, command)
       end
     end
 
@@ -72,7 +72,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
       create_hash = { name: name, ensure: 'create' }
       array_of_commands_to_run = Puppet::Provider::NetworkVlan::CiscoIos.commands_from_instance(create_hash)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_conf_t_mode(command)
+        context.transport.run_command_conf_t_mode(command)
       end
       update(context, name, should)
     end
@@ -81,7 +81,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
       clear_hash = { name: name, ensure: 'absent' }
       array_of_commands_to_run = Puppet::Provider::NetworkVlan::CiscoIos.commands_from_instance(clear_hash)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_conf_t_mode(command)
+        context.transport.run_command_conf_t_mode(command)
       end
     end
