+++ lib/puppet/provider/ntp_auth_key/cisco_ios.rb
@@ -38,7 +38,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     end
 
     def get(context, _names = nil)
-      output = context.device.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
+      output = context.transport.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
       return [] if output.nil?
       return_value = Puppet::Provider::NtpAuthKey::CiscoIos.instances_from_cli(output)
       PuppetX::CiscoIOS::Utility.enforce_simple_types(context, return_value)
@@ -47,7 +47,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     def create(context, _name, should)
       array_of_commands_to_run = Puppet::Provider::NtpAuthKey::CiscoIos.commands_from_instance(should)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_conf_t_mode(command)
+        context.transport.run_command_conf_t_mode(command)
       end
     end
 
@@ -57,7 +57,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
       clear_hash = { name: name, ensure: 'absent' }
       array_of_commands_to_run = Puppet::Provider::NtpAuthKey::CiscoIos.commands_from_instance(clear_hash)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_conf_t_mode(command)
+        context.transport.run_command_conf_t_mode(command)
       end
     end
