+++ lib/puppet/provider/ntp_config/cisco_ios.rb
@@ -46,7 +46,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     end
 
     def get(context, _names = nil)
-      output = context.device.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
+      output = context.transport.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
       return [] if output.nil?
       return_value = Puppet::Provider::NtpConfig::CiscoIos.instances_from_cli(output)
       PuppetX::CiscoIOS::Utility.enforce_simple_types(context, return_value)
@@ -66,7 +66,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     def update(context, _name, is, should)
       array_of_commands_to_run = Puppet::Provider::NtpConfig::CiscoIos.commands_from_is_should(is, should)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_conf_t_mode(command)
+        context.transport.run_command_conf_t_mode(command)
       end
     end
