+++ lib/puppet/provider/radius_server_group/cisco_ios.rb
@@ -46,7 +46,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     end
 
     def get(context, _names = nil)
-      output = context.device.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
+      output = context.transport.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
       return [] if output.nil?
       return_value = Puppet::Provider::RadiusServerGroup::CiscoIos.instances_from_cli(output)
       PuppetX::CiscoIOS::Utility.enforce_simple_types(context, return_value)
@@ -82,7 +82,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
       # only called if adding/removing servers
       array_of_commands_to_run = Puppet::Provider::RadiusServerGroup::CiscoIos.commands_from_is_should(is, should)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_radius_mode(name, command)
+        context.transport.run_command_radius_mode(name, command)
       end
     end
 
@@ -90,14 +90,14 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
       clear_hash = { name: name, ensure: 'absent' }
       array_of_commands_to_run = Puppet::Provider::RadiusServerGroup::CiscoIos.commands_from_instance(clear_hash)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_conf_t_mode(command)
+        context.transport.run_command_conf_t_mode(command)
       end
     end
 
     def create(context, name, should)
       array_of_commands_to_run = Puppet::Provider::RadiusServerGroup::CiscoIos.commands_from_instance(should)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_conf_t_mode(command)
+        context.transport.run_command_conf_t_mode(command)
       end
       is = { name: name, ensure: 'absent' }
       update(context, name, is, should)
