+++ lib/puppet/provider/syslog_server/cisco_ios.rb
@@ -40,7 +40,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     end
 
     def get(context, _names = nil)
-      output = context.device.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
+      output = context.transport.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
       return [] if output.nil?
       Puppet::Provider::SyslogServer::CiscoIos.instances_from_cli(output)
     end
@@ -48,7 +48,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     def update(context, _name, should)
       array_of_commands_to_run = Puppet::Provider::SyslogServer::CiscoIos.commands_from_instance(should)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_conf_t_mode(command)
+        context.transport.run_command_conf_t_mode(command)
       end
     end
 
@@ -58,7 +58,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
       clear_hash = { name: name, ensure: 'absent' }
       array_of_commands_to_run = Puppet::Provider::SyslogServer::CiscoIos.commands_from_instance(clear_hash)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_conf_t_mode(command)
+        context.transport.run_command_conf_t_mode(command)
       end
     end
