+++ lib/puppet/provider/tacacs_server_group/cisco_ios.rb
@@ -46,7 +46,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     end
 
     def get(context, _names = nil)
-      output = context.device.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
+      output = context.transport.run_command_enable_mode(PuppetX::CiscoIOS::Utility.get_values(commands_hash))
       return [] if output.nil?
       return_value = Puppet::Provider::TacacsServerGroup::CiscoIos.instances_from_cli(output)
       PuppetX::CiscoIOS::Utility.enforce_simple_types(context, return_value)
@@ -81,7 +81,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
     def update(context, _name, is, should)
       array_of_commands_to_run = Puppet::Provider::TacacsServerGroup::CiscoIos.commands_from_is_should(is, should)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_tacacs_server_group_mode(should[:name], command)
+        context.transport.run_command_tacacs_server_group_mode(should[:name], command)
       end
     end
 
@@ -89,7 +89,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
       delete_hash = { name: name, ensure: 'absent' }
       array_of_commands_to_run = Puppet::Provider::TacacsServerGroup::CiscoIos.commands_from_instance(delete_hash)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_conf_t_mode(command)
+        context.transport.run_command_conf_t_mode(command)
       end
     end
 
@@ -98,7 +98,7 @@ unless PuppetX::CiscoIOS::Check.use_old_netdev_type
       instance = should.clone
       array_of_commands_to_run = Puppet::Provider::TacacsServerGroup::CiscoIos.commands_from_instance(instance)
       array_of_commands_to_run.each do |command|
-        context.device.run_command_conf_t_mode(command)
+        context.transport.run_command_conf_t_mode(command)
       end
       is = { name: name, ensure: 'present' }
       update(context, name, is, should)
